rename to catenate
